updating header descriptions and moved pointer function also added description for the pointer function
